+++ mex/drop/api.py
@@ -6,6 +6,7 @@ from fastapi import (
     Body,
     Depends,
     File,
+    Header,
     HTTPException,
     Path,
     Response,
@@ -15,7 +16,12 @@ from pydantic import BaseModel
 from starlette import status
 from starlette.background import BackgroundTask, BackgroundTasks
 
-from mex.common.exceptions import MExError
+from mex.drop.files_io import (
+    ALLOWED_CONTENT_TYPES,
+    check_duplicate_filenames,
+    validate_file_extension,
+    write_to_file,
+)
 from mex.drop.security import get_current_authorized_x_systems, is_authorized
 from mex.drop.settings import DropSettings
 from mex.drop.sinks.json import json_sink
@@ -26,7 +32,7 @@ router = APIRouter(prefix="/v0", tags=["api"])
 
 @router.post(
     "/{x_system}/{entity_type}",
-    description="Upload arbitrary JSON data to MEx.",
+    description="Upload arbitrary structured data to MEx.",
     tags=["API"],
     status_code=202,
 )
@@ -45,15 +51,15 @@ async def drop_data(
             default=...,
             pattern=PATH_REGEX,
             description=(
-                "Name of the data file that is uploaded, " "if unsure use 'default'"
+                "Name of the data file that is uploaded, if unsure use 'default'"
             ),
         ),
     ],
     data: Annotated[
-        dict[str, Any] | list[Any],
+        dict[str, Any] | list[Any] | bytes,
         Body(
             description=(
-                "An arbitrary JSON structure, " "that can be further processed by MEx"
+                "An arbitrary data structure, that can be further processed by MEx"
             ),
             examples=[
                 {"foo": "bar", "list": [1, 2, "foo"], "nested": {"foo": "bar"}},
@@ -61,17 +67,27 @@ async def drop_data(
             ],
         ),
     ],
+    content_type: Annotated[
+        str,
+        Header(
+            description=(
+                f"Content-Type of the uploaded data.\n Allowed content types:"
+                f"\n{', '.join(ALLOWED_CONTENT_TYPES.keys())}"
+            ),
+        ),
+    ],
     authorized_x_systems: Annotated[
         list[XSystem], Depends(get_current_authorized_x_systems)
     ],
 ) -> Response:
-    """Upload arbitrary JSON data to MEx.
+    """Upload arbitrary data to MEx.
 
     Args:
         x_system: name of the x-system that the data comes from
         entity_type: name of the data file that is uploaded, if unsure use 'default'
-        data: dictionary with string keys or list with and arbitrary values
+        data: data content of request body
         authorized_x_systems: list of authorized x-systems
+        content_type: Content-Type of the uploaded data
 
     Settings:
         drop_directory: where accepted data is stored
@@ -84,10 +100,26 @@ async def drop_data(
             status_code=status.HTTP_403_FORBIDDEN,
             detail="API Key not authorized to drop data for this x_system.",
         )
+    if content_type not in ALLOWED_CONTENT_TYPES:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail=f"Unsupported Content-Type. "
+            f"Allowed types: {', '.join(ALLOWED_CONTENT_TYPES.keys())}",
+        )
+    file_ext = ALLOWED_CONTENT_TYPES[content_type]
     settings = DropSettings.get()
-    out_file = pathlib.Path(settings.drop_directory, x_system, entity_type + ".json")
-    return Response(
-        status_code=202, background=BackgroundTask(json_sink, data, out_file)
+    out_file = settings.drop_directory / x_system / f"{entity_type}{file_ext}"
+    if content_type == "application/json" and isinstance(data, (dict | list)):
+        return Response(
+            status_code=200, background=BackgroundTask(json_sink, data, out_file)
+        )
+    if isinstance(data, bytes):
+        return Response(
+            status_code=200, background=BackgroundTask(write_to_file, data, out_file)
+        )
+    raise HTTPException(
+        status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
+        detail="Unsupported content typeor format.",
     )
 
 
@@ -103,13 +135,16 @@ async def drop_data_multipart(
         Path(
             default=...,
             pattern=PATH_REGEX,
-            description="Name of the system that the data comes from",
+            description="Name of the system that the data comes from.",
         ),
     ],
     files: Annotated[
         list[UploadFile],
         File(
-            description=("Multipart file list, " "that can be further processed by MEx")
+            description=(
+                f"Multipart file list, that can be further processed by MEx.\n "
+                f"Allowed file types: {', '.join(ALLOWED_CONTENT_TYPES.values())}"
+            )
         ),
     ],
     authorized_x_systems: Annotated[
@@ -121,7 +156,6 @@ async def drop_data_multipart(
 
     Args:
         x_system: name of the x-system that the data comes from
-        entity_type: name of the data file that is uploaded, if unsure use 'default'
         files: list of files to be uploaded to MEx
         authorized_x_systems: list of authorized x-systems
         background_tasks: collection of background tasks
@@ -138,24 +172,14 @@ async def drop_data_multipart(
             detail="API Key not authorized to drop data for this x_system.",
         )
     settings = DropSettings.get()
+    check_duplicate_filenames(files)
     for file in files:
-        content = await file.read()
         entity_type = str(file.filename)
+        await validate_file_extension(file.content_type, entity_type)
+        content = await file.read()
         out_file = pathlib.Path(settings.drop_directory, x_system, entity_type)
         background_tasks.add_task(write_to_file, content, out_file)
     return Response(status_code=202)
-
-
-async def write_to_file(content: bytes, out_file: pathlib.Path) -> None:
-    """Write content to file."""
-    out_file.parent.mkdir(parents=True, exist_ok=True)
-    try:
-        with open(out_file, "wb") as f:
-            f.write(content)
-    except Exception as exc:
-        raise MExError(
-            f"Failed to write to file {out_file}: {exc!s}",
-        ) from exc
 
 
 @router.get(