+++ tests/test_api.py
@@ -4,14 +4,17 @@ from pathlib import Path
 from typing import Any
 from unittest.mock import AsyncMock, call
 
+import pandas as pd
 import pytest
 from fastapi.testclient import TestClient
 from pytest import MonkeyPatch
 
 import mex
-import mex.drop.api
+from mex.drop.files_io import ALLOWED_CONTENT_TYPES
 from mex.drop.settings import DropSettings
 from mex.drop.types import EntityType, XSystem
+
+TESTDATA_DIR = Path(__file__).parent / "test_files"
 
 
 @pytest.fixture
@@ -34,19 +37,138 @@ def dropped_data(tmp_path: Path) -> dict[str, Any]:
 
 
 @pytest.mark.parametrize(
-    "api_key, x_system, entity_type, expected_response_code",
+    (
+        "api_key",
+        "x_system",
+        "entity_type",
+        "expected_response_code",
+        "content_type",
+        "expected_content",
+    ),
     [
-        ("api-test-key", "test_system", "valid_entity_type", 202),
-        ("api-test-key", "foo_system", "valid_entity_type", 202),
-        ("api-test-key", "test_system", "invalid entity type", 422),
-        ("api-test-key", "invalid x_system", "valid_entity_type", 422),
-        (None, "test_system", "valid_entity_type", 401),
-        ("invalid-key", "test_system", "valid_entity_type", 401),
-        ("api-key-one", "foo_system", "valid_entity_type", 403),
+        (
+            "api-test-key",
+            "test_system",
+            "valid_entity_type",
+            200,
+            "application/json",
+            {
+                "asd": "def",
+                "foo": 1,
+                "bar": 1.2,
+                "list": [1, 2, 3],
+                "dict": {"a": "b"},
+            },
+        ),
+        (
+            "api-test-key",
+            "foo_system",
+            "valid_entity_type",
+            200,
+            "text/csv",
+            "asd,foo,bar,list,dict.a\ndef,1,1.2,\"[1, 2, 3]\",\"{'a': 'b'}\"\n",
+        ),
+        (
+            "api-test-key",
+            "test_system",
+            "valid_entity_type",
+            200,
+            "application/xml",
+            (TESTDATA_DIR / "test.xml").read_text(),
+        ),
+        (
+            "api-test-key",
+            "test_system",
+            "valid_entity_type",
+            200,
+            "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
+            (TESTDATA_DIR / "test.xlsx").read_bytes(),
+        ),
+        (
+            "api-test-key",
+            "test_system",
+            "valid_entity_type",
+            400,
+            "application/pdf",
+            "foo",
+        ),
+        (
+            "api-test-key",
+            "test_system",
+            "invalid entity type",
+            422,
+            "application/json",
+            {
+                "asd": "def",
+                "foo": 1,
+                "bar": 1.2,
+                "list": [1, 2, 3],
+                "dict": {"a": "b"},
+            },
+        ),
+        (
+            "api-test-key",
+            "invalid x_system",
+            "valid_entity_type",
+            422,
+            "application/json",
+            {
+                "asd": "def",
+                "foo": 1,
+                "bar": 1.2,
+                "list": [1, 2, 3],
+                "dict": {"a": "b"},
+            },
+        ),
+        (
+            None,
+            "test_system",
+            "valid_entity_type",
+            401,
+            "application/json",
+            {
+                "asd": "def",
+                "foo": 1,
+                "bar": 1.2,
+                "list": [1, 2, 3],
+                "dict": {"a": "b"},
+            },
+        ),
+        (
+            "invalid-key",
+            "test_system",
+            "valid_entity_type",
+            401,
+            "application/json",
+            {
+                "asd": "def",
+                "foo": 1,
+                "bar": 1.2,
+                "list": [1, 2, 3],
+                "dict": {"a": "b"},
+            },
+        ),
+        (
+            "api-key-one",
+            "foo_system",
+            "valid_entity_type",
+            403,
+            "application/json",
+            {
+                "asd": "def",
+                "foo": 1,
+                "bar": 1.2,
+                "list": [1, 2, 3],
+                "dict": {"a": "b"},
+            },
+        ),
     ],
     ids=(
         "valid",
-        "valid2",
+        "valid csv",
+        "valid xml",
+        "valid xlsx",
+        "invalid content type",
         "invalid entity type",
         "invalid x_system",
         "missing header",
@@ -60,51 +182,123 @@ def test_drop_data(
     x_system: XSystem,
     entity_type: EntityType,
     expected_response_code: int,
+    content_type: str,
+    expected_content: Any,
     monkeypatch: MonkeyPatch,
     settings: DropSettings,
 ) -> None:
     mocked_sink = AsyncMock(return_value=None)
     monkeypatch.setattr(mex.drop.api, "json_sink", mocked_sink)
-    expected_content = {
-        "asd": "def",
-        "foo": 1,
-        "bar": 1.2,
-        "list": [1, 2, 3],
-        "dict": {"a": "b"},
-    }
-    expected_file = Path(settings.drop_directory, x_system, entity_type + ".json")
-    response = client.post(
-        f"/v0/{x_system}/{entity_type}",
-        headers={"X-API-Key": api_key} if api_key else {},
-        json=expected_content,
+
+    if api_key:
+        client.headers.update({"X-API-Key": api_key})
+    client.headers.update({"Content-Type": content_type})
+    kwargs = (
+        {"json": expected_content}
+        if content_type == "application/json"
+        else {"content": expected_content}
     )
+    response = client.post(f"/v0/{x_system}/{entity_type}", **kwargs)
     assert response.status_code == expected_response_code, response.text
+
     if 200 <= response.status_code < 300:
-        assert mocked_sink.call_args == call(expected_content, expected_file)
+        base_path = Path(settings.drop_directory, x_system, entity_type)
+        expected_file = base_path.with_suffix(ALLOWED_CONTENT_TYPES[content_type])
+
+        if content_type == "application/json":
+            assert mocked_sink.call_args == call(expected_content, expected_file)
+        else:
+            with open(expected_file, "rb") as f:
+                saved_content = f.read()
+
+            if (
+                content_type == "text/csv"
+                or content_type == "text/tab-separated-values"
+            ):
+                assert saved_content.decode("utf-8") == expected_content
+            elif (
+                content_type
+                == "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
+                or content_type == "application/vnd.ms-excel"
+            ):
+                original_df = pd.read_excel(BytesIO(expected_content), sheet_name=None)
+                saved_df = pd.read_excel(BytesIO(saved_content), sheet_name=None)
+
+                assert original_df.keys() == saved_df.keys()
+
+                for sheet in original_df.keys():
+                    pd.testing.assert_frame_equal(original_df[sheet], saved_df[sheet])
 
 
 @pytest.mark.parametrize(
-    "api_key, x_system, expected_response_code, files",
+    ("api_key", "x_system", "expected_response_code", "files"),
     [
         (
             "api-test-key",
             "test_system",
             202,
-            {"file1.txt": "file1 content", "file2.csv": "1,2,3"},
+            [
+                ("file1.json", "file1 content", "application/json"),
+                ("file2.csv", "1,2,3", "text/csv"),
+            ],
+        ),
+        (
+            "api-test-key",
+            "test_system",
+            202,
+            [("file1.xls", "file1 content", "application/vnd.ms-excel")],
+        ),
+        (
+            "api-test-key",
+            "test_system",
+            422,
+            [("file1.html", "file1 content", "text/html")],
         ),
         ("api-test-key", "foo_system", 422, {}),
-        ("api-test-key", "invalid x_system", 422, {"file1.txt": "file1 content"}),
-        (None, "test_system", 401, {"file1.txt": "file1 content"}),
-        ("invalid-key", "test_system", 401, {"file1.txt": "file1 content"}),
-        ("api-key-one", "foo_system", 403, {"file1.txt": "file1 content"}),
+        (
+            "api-test-key",
+            "invalid x_system",
+            422,
+            [("file1.json", "file1 content", "application/json")],
+        ),
+        (
+            None,
+            "test_system",
+            401,
+            [("file1.json", "file1 content", "application/json")],
+        ),
+        (
+            "invalid-key",
+            "test_system",
+            401,
+            [("file1.json", "file1 content", "application/json")],
+        ),
+        (
+            "api-key-one",
+            "foo_system",
+            403,
+            [("file1.json", "file1 content", "application/json")],
+        ),
+        (
+            "api-test-key",
+            "test_system",
+            400,
+            [
+                ("file1.json", "file1 content", "application/json"),
+                ("file1.json", "file1 different content", "application/json"),
+            ],
+        ),
     ],
     ids=(
         "valid",
+        "valid xls",
+        "invalid file format",
         "missing upload",
         "invalid x_system",
         "missing header",
         "invalid api_key",
         "unauthorized x_system",
+        "duplicate filename",
     ),
 )
 def test_drop_multiple_files(
@@ -112,12 +306,12 @@ def test_drop_multiple_files(
     api_key: str | None,
     x_system: XSystem,
     expected_response_code: int,
-    files: dict[str, str],
+    files: list[tuple[str, str, str]],
     settings: DropSettings,
 ) -> None:
     files_data = [
-        ("files", (name, BytesIO(content.encode()), "text/plain"))
-        for name, content in files.items()
+        ("files", (name, BytesIO(content.encode()), content_type))
+        for name, content, content_type in files
     ]
     response = client.post(
         f"/v0/{x_system}",
@@ -126,8 +320,8 @@ def test_drop_multiple_files(
     )
     assert response.status_code == expected_response_code, response.text
     if 200 <= response.status_code < 300 and files:
-        for filename, content in files.items():
-            expected_file1 = Path(settings.drop_directory, x_system, filename)
+        for name, content, _ in files:
+            expected_file1 = Path(settings.drop_directory, x_system, name)
             assert expected_file1.read_text() == content