+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- support for structured file formats (xml, tsv, xls and xlsx)
+
 ### Changes
 
 ### Deprecated